Links to landing pages should be via visible URLs
The SDP search results page shows resource titles as links that link to
the resource's PDR landing page.  These links, however, are currently
implemented as a javascript operation rather than as a traditional HTML
link (as in, {{<a href="/od/id/...">}}).  Consequently, this prevents
link interactions that users are used to; in particular, they cannot:
* See the URL displayed at the bottom of their browser when they hover
over the link
* Copy the URL into their paste buffer (via right click -> "Copy Link
Location")
* Load page into a new window or tab (via right click -> "Open in new
window|tab")

This link should be implemented with a traditional HTML {{<a
href="...">}} element where the {{href}} value is set to a real URL.