UI changes
UI fixes based on Bob's feedback
o  The font in the menu bar seems a bit heavy to me.  The previous
version is lighter, cleaner, and easier to read.
o  I prefer to see "NIST Data Discovery" rather than "NIST DATA
DISCOVERY" because it is clearer that NIST is an acronym.
o  And like above, the font used for "Explore data, tools, ..." looks
heavier than before.