advanced search changes
Fixed -

1) When deleting a query on the left panel, there is no confirmation
dialog. Please add the confirmation dialog you  have on the right side
panel

2) When saving an existing query with a new name, the old is still
present along with the new. As a user I should be able to rename my
query. If I am saving a query I should be able to provide a new name.
The tool should not create a new query with the new name but instead
should just change the name of the existing one.

3) Export button should be disable when no query is present.

4) Export and Import buttons are missing their icons.